[PyTorch] Move NEON VecConvert specialization from vec256_convert to vec128_convert
NEON vectors are 128-bit and don't belong with 256 stuff.

Differential Revision: [D64143615](https://our.internmc.facebook.com/intern/diff/D64143615/)

[ghstack-poisoned]
